test(libbridge): split link-resume by behaviour family
Split link-resume.test.js (480 LOC) into link-resume-prepare.test.js
(prepareLinkResume) and link-resume-complete.test.js
(createLinkCompleteHandler) along the two top-level describe boundaries.
Shared TRUSTED/SECRET/NOW/clock constants lifted into
link-resume-helpers.js. Assertions unchanged; 22 tests preserved.

callback-handler, dispatcher and resume-scheduler are retained over the
ceiling on the allow-list — each is one cohesive single-function
behaviour with no clean family seam.

Spec 0640 § C (large files), design Decision 5.
