+++ libraries/libbridge/test/link-resume-complete.test.js
@@ -1,102 +1,12 @@
 import { describe, expect, test } from "bun:test";
 import { Hono } from "hono";
 
-import {
-  prepareLinkResume,
-  createLinkCompleteHandler,
-} from "../src/link-resume.js";
+import { createLinkCompleteHandler } from "../src/link-resume.js";
 import {
   mintCompletionTicket,
   TICKET_TTL_MS,
 } from "@forwardimpact/libutil/completion-ticket";
-import { loadTrustedIdpOrigins } from "@forwardimpact/libutil/trusted-origins";
-
-const TRUSTED = loadTrustedIdpOrigins("https://oauth.example");
-const SECRET = "test-secret-aaaaaaaaaaaaaaaaaaaaaaaaaaaa";
-const NOW = 1_700_000_000_000;
-const clock = { now: () => NOW };
-
-describe("prepareLinkResume — keyword args, discriminated return", () => {
-  test("returns { linkToken, augmentedUrl } for a trusted, parseable URL", () => {
-    const r = prepareLinkResume({
-      authorizeUrl:
-        "https://oauth.example/authorize?surface=github-discussions&surface_user_id=42",
-      callbackBaseUrl: "https://bridge.example/",
-      trustedOrigins: TRUSTED,
-    });
-    expect(r.skipped).toBeUndefined();
-    const url = new URL(r.augmentedUrl);
-    expect(url.searchParams.get("redirect_uri")).toBe(
-      "https://bridge.example/api/link-complete",
-    );
-    expect(url.searchParams.get("client_state")).toBe(r.linkToken);
-    expect(typeof r.linkToken).toBe("string");
-    expect(r.linkToken.length).toBeGreaterThan(0);
-  });
-
-  test("produces unique tokens on successive calls", () => {
-    const a = prepareLinkResume({
-      authorizeUrl: "https://oauth.example/a",
-      callbackBaseUrl: "https://b",
-      trustedOrigins: TRUSTED,
-    });
-    const b = prepareLinkResume({
-      authorizeUrl: "https://oauth.example/a",
-      callbackBaseUrl: "https://b",
-      trustedOrigins: TRUSTED,
-    });
-    expect(a.linkToken).not.toBe(b.linkToken);
-  });
-
-  test("strips trailing slash from callbackBaseUrl", () => {
-    const { augmentedUrl } = prepareLinkResume({
-      authorizeUrl: "https://oauth.example/authorize",
-      callbackBaseUrl: "https://bridge.example///",
-      trustedOrigins: TRUSTED,
-    });
-    const url = new URL(augmentedUrl);
-    expect(url.searchParams.get("redirect_uri")).toBe(
-      "https://bridge.example/api/link-complete",
-    );
-  });
-
-  test("untrusted origin → { skipped, reason: 'untrusted_origin' }", () => {
-    const r = prepareLinkResume({
-      authorizeUrl: "https://attacker.example/login",
-      callbackBaseUrl: "https://bridge.example",
-      trustedOrigins: TRUSTED,
-    });
-    expect(r).toEqual({ skipped: true, reason: "untrusted_origin" });
-  });
-
-  test("malformed URL → { skipped, reason: 'untrusted_origin' }", () => {
-    const r = prepareLinkResume({
-      authorizeUrl: "not-a-url",
-      callbackBaseUrl: "https://bridge.example",
-      trustedOrigins: TRUSTED,
-    });
-    expect(r).toEqual({ skipped: true, reason: "untrusted_origin" });
-  });
-
-  test("missing trustedOrigins throws TypeError (forget-resistance)", () => {
-    expect(() =>
-      prepareLinkResume({
-        authorizeUrl: "https://oauth.example/a",
-        callbackBaseUrl: "https://b",
-      }),
-    ).toThrow(TypeError);
-  });
-
-  test("non-Set trustedOrigins throws TypeError", () => {
-    expect(() =>
-      prepareLinkResume({
-        authorizeUrl: "https://oauth.example/a",
-        callbackBaseUrl: "https://b",
-        trustedOrigins: ["https://oauth.example"],
-      }),
-    ).toThrow(TypeError);
-  });
-});
+import { TRUSTED, SECRET, NOW, clock } from "./link-resume-helpers.js";
 
 describe("createLinkCompleteHandler — ticket verification gates store touch", () => {
   function makeApp(handler) {
